refactor : (로또 게임) 로또게임머신 클래스 책임 분리
